[libc++] [test] Rename deduct.fail.cpp -> deduct.verify.cpp, and reformat
All tests have been reformatted to 2-space indents and commented
in the modern style. Remove the repetition of Standardese deduction guides,
since the important thing about these verify tests is that they reject
the right things, not that they "exercise" specific signatures (which
by definition they will not, because the point is that no specific signature
matches, hence CTAD fails).
